Convert the easing helpers to typescript
