--- types/helpers/helpers.easing.d.ts
@@ -1,5 +1,0 @@
-import { EasingFunction } from '..';
-
-export type EasingFunctionSignature = (t: number) => number;
-
-export declare const easingEffects: Record<EasingFunction, EasingFunctionSignature>;
